executed all test cases so now we have outputfiles

+++ build.sh
@@ -1,6 +1,8 @@
 if [ ! -d "bin" ]; then
     mkdir bin
 else
-	rm bin/*
+    if [ "$(ls -A bin)" ]; then
+	    rm bin/*
+    fi
 fi
 g++ -g -O0 -I . -o bin/interrupts interrupts.cpp
